[CodeCompletion] Extract CodeCompletionStringPrinter.h to its own file
